+++ packages/gestalt/src/TapArea.js
@@ -61,7 +61,12 @@ function TapArea({
   tapStyle = 'none',
   rounding = 0,
 }: Props) {
+  const innerRef = React.useRef(null);
+  // $FlowFixMe Flow thinks forwardedRef is a number, which is incorrect
+  React.useImperativeHandle(forwardedRef, () => innerRef.current);
+
   const {
+    compressStyle,
     isTapping,
     handleBlur,
     handleMouseDown,
@@ -70,9 +75,13 @@ function TapArea({
     handleTouchMove,
     handleTouchCancel,
     handleTouchEnd,
-  } = useTapFeedback();
+  } = useTapFeedback({
+    height: innerRef?.current?.clientHeight,
+    width: innerRef?.current?.clientWidth,
+  });
 
   const className = classnames(
+    styles.tapTransition,
     styles.touchable,
     getRoundingClassName(rounding),
     {
@@ -133,8 +142,11 @@ function TapArea({
       onTouchMove={handleTouchMove}
       onTouchCancel={handleTouchCancel}
       onTouchEnd={handleTouchEnd}
-      ref={forwardedRef}
+      ref={innerRef}
       role="button"
+      {...(compressStyle && tapStyle === 'compress'
+        ? { style: compressStyle }
+        : {})}
       tabIndex={disabled ? null : '0'}
     >
       {children}